sow-backend/requirements.txt
fastapi
uvicorn[standard]
gunicorn
python-dotenv
pydantic
requests
pytest
pydantic-settings
python-dotenv
python-docx
PyPDF2
openai
regex
debugpy
azure-storage-blob
httpx
psycopg2-binary
azure-storage-blob
python-multipart
reportlab
weasyprint
python-jose[cryptography]
passlib[bcrypt]
bcrypt==4.0.1
email-validator